Aplicando classes de base e extensões nas classes de permissão

+++ src/repositories/prisma/adm/prisma-permissions.repositories.ts
@@ -1,45 +1,19 @@
 import { PermissionsDb } from '@prisma/client';
-import { Permission } from '../../../modules/adm/permissions/schema';
 import { IPermissionsRepositories } from '../../adm/permissions.repositories';
+import { PrismaBaseRepositories } from '../base/prisma-base.repositories';
 import { prisma } from '../../../database/prisma';
 
-export class PrismaPermissionsRepositories implements IPermissionsRepositories {
+export class PrismaPermissionsRepositories extends PrismaBaseRepositories<PermissionsDb, string> implements IPermissionsRepositories {
 
-    private readonly db = prisma.permissionsDb;
-
-    async findById(id: string): Promise<PermissionsDb | null> {
-        const data = await this.db.findUnique({ where: { id } });
-        return data;
+    constructor() {
+        super(prisma, (client) => client.permissionsDb);
     }
 
     async findByRoleId(roleId: string): Promise<PermissionsDb[]> {
-        const data = await this.db.findMany({ where: { roleId } });
-        return data;
-    }
-
-    async save(obj: Permission): Promise<PermissionsDb> {
-        const newObj = await this.db.create({
-            data: { ...obj }
-        });
-
-        return newObj;
-    }
-
-    async update(obj: Permission): Promise<PermissionsDb> {
-        const id = obj.id;
-        const newObj = await this.db.update({
-            where: { id },
-            data: { ...obj }
-        });
-
-        return newObj;
-    }
-
-    async delete(id: string): Promise<void> {
-        await this.db.delete({ where: { id } });
+        return await this.db.permissionsDb.findMany({ where: { roleId } });
     }
 
     async deleteByRoleId(roleId: string): Promise<void> {
-        await this.db.deleteMany({ where: { roleId } });
+        await this.db.permissionsDb.deleteMany({ where: { roleId } });
     }
 }
